+++ src/graphics/graphics/vulkan/Vertex.hh
@@ -45,7 +45,7 @@ namespace sp::vulkan {
         vk::PipelineVertexInputStateCreateInfo pipelineInputInfo;
         std::array<vk::VertexInputBindingDescription, MAX_VERTEX_INPUT_BINDINGS> bindings;
         std::array<vk::VertexInputAttributeDescription, MAX_VERTEX_ATTRIBUTES> attributes;
-        size_t bindingCount = 0, attributeCount = 0;
+        uint32_t bindingCount = 0, attributeCount = 0;
     };
 
     struct TextureVertex {
